+++ app/views/AddNewSourceDatasetView/common/entities.ts
@@ -1,0 +1,4 @@
+import { InferType } from "yup";
+import { newSourceDatasetSchema } from "./schema";
+
+export type NewSourceDatasetData = InferType<typeof newSourceDatasetSchema>;
